+++ tests/listbox.test.js
@@ -44,8 +44,8 @@ test("Should properly check items in the listbox", async () => {
     const options = getOptions()
     assert.equal(options[index].selected, false)
     assert.equal(options[index].current, false)
-    assert.equal(options[index].getAttribute("aria-selected"), "false")
-    assert.equal(options[index].getAttribute("aria-current"), "false")
+    assert.isFalse(options[index].hasAttribute("aria-selected"))
+    assert.isFalse(options[index].hasAttribute("aria-current"))
     // assert.equal(options[index].hasAttribute("aria-selected"), false)
     // assert.equal(options[index].hasAttribute("aria-current"), false)
   }
@@ -107,7 +107,7 @@ test("Should properly set aria-selected and aria-checked for options in a multis
   const isNotSelected = (index) => {
     const options = getOptions()
     assert.equal(options[index].selected, false)
-    assert.equal(options[index].getAttribute("aria-selected"), "false")
+    assert.isFalse(options[index].hasAttribute("aria-selected"))
     // assert.equal(options[index].hasAttribute("aria-selected"), false)
   }
 
@@ -121,7 +121,7 @@ test("Should properly set aria-selected and aria-checked for options in a multis
   const isNotCurrent = (index) => {
     const options = getOptions()
     assert.equal(options[index].current, false)
-    assert.equal(options[index].getAttribute("aria-current"), "false")
+    assert.isFalse(options[index].hasAttribute("aria-current"))
     // assert.equal(options[index].hasAttribute("aria-current"), false)
   }