+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-# v3.0.1 10/01/2024
+## v3.0.2 10/17/2024
+
+- Multiple bug fixes in `<role-combobox>`
+- Added the `hide-selected-options` for hiding options from the list that have already been selected.
+
+## v3.0.1 10/01/2024
 
 - Fixed a bug if interactive elements were children of `<role-toolbar>`
 
-# v3.0.0 09/21/2024
+## v3.0.0 09/21/2024
 
 - Removed `role-slider`
 - Fixed a bug in `<role-combobox>`